# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AquaForge API",
    "description": "Swim Meet Optimization API - Backend services for AquaForge.ai",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/export": {
      "post": {
        "tags": [
          "Export"
        ],
        "summary": "Export Results",
        "description": "Export optimization results to the specified format.\n\nArgs:\n    request: Export request with format and results data\n\nReturns:\n    Export response with content or download URL",
        "operationId": "export_results_api_v1_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/download/{filename}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Download Export",
        "description": "Download a previously generated export file.\n\nNote: This requires server-side file storage to be implemented.",
        "operationId": "download_export_api_v1_export_download__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response."
      },
...
      "ExportFormat": {
        "type": "string",
        "enum": [
          "csv",
          "pdf",
          "html",
          "json"
        ],
        "title": "ExportFormat",
        "description": "Supported export formats."
      },
      "ExportRequest": {
        "properties": {
          "format": {
            "$ref": "#/components/schemas/ExportFormat",
            "description": "Export format"
          },
          "optimization_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Optimization Results",
            "description": "Optimization results to export"
          },
          "seton_score": {
            "type": "number",
            "title": "Seton Score",
            "default": 0.0
          },
          "opponent_score": {
            "type": "number",
            "title": "Opponent Score",
            "default": 0.0
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "format",
          "optimization_results"
        ],
        "title": "ExportRequest",
        "description": "Request model for exporting results."
      },
      "ExportResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "Base64 encoded content for binary formats"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "URL for file download if stored"
          }
        },
        "type": "object",
        "required": [
          "success",
          "format",
          "filename"
        ],
        "title": "ExportResponse",
        "description": "Response from export request."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}